# openapi.yaml
...
openapi: 3.1.0
info:
  title: Silicon Hook API
  version: 0.5.0
  summary: Signed webhook endpoints with ordered, acknowledged delivery to Silicons.
  description: |
    Silicon Hook gives each Silicon signed webhook endpoints at
    `https://hook.teamofsilicons.com/silicon/{silicon_id}/{endpoint_key}`.
    Every request to an active endpoint is answered with `webhook.ok`. Requests
    that verify against the hook's signature policy are retained for 14 days
    and delivered in order over the WebSocket stream at `/api/v1/ws` or by
    polling `/api/v1/silicons/{silicon_id}/deliveries`; requests that do not
    verify are withheld and kept in a separate 14-day blocked log. Timestamps
    are UTC RFC 3339 values. Endpoint keys route requests but never
    authenticate them. IAM test app_secret selectors choose isolated sandboxes via
    X-Hook-Test-App-Secret; they never grant actor authority. Deprecated API v1
    contracts sunset only after seven request-free days. Contract accounting is
    independent of optional diagnostic telemetry.
  contact:
    email: engineering@teamofsilicons.com
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
- url: https://backend.hook.teamofsilicons.com
  description: Backend API; public ingress may use the configured Hook frontend origin.
security:
- bearerAuth: []
...
paths:
...
  "/api/v1/testing-environments":
    post:
      tags:
      - Testing
      operationId: createTestingEnvironment
      summary: Create an empty sandbox linked to a real IAM test environment
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/IdempotencyKey"
      responses:
        '200':
          description: Operation completed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EnvironmentWithKey"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateEnvironment"
    get:
      tags:
      - Testing
      operationId: listTestingEnvironments
      summary: List this organization’s environments
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - name: status
        in: query
        schema:
          type: string
          enum:
          - active
          - deleted
          - all
          default: active
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100
      - name: after
        in: query
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Operation completed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EnvironmentList"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
...
  "/api/v1/testing-environment":
    get:
      tags:
      - Testing
      operationId: currentTestingEnvironment
      summary: Read root metadata using the Hook test key
      parameters:
      - &1
        name: X-Hook-Test-Key
        in: header
        required: true
        description: Selects an isolated Hook environment. The bearer must be from
          its linked IAM test world. Never attach to public ingress.
        schema:
          type: string
          pattern: "^[A-Za-z0-9]{32}$"
      responses:
        '200':
          description: Operation completed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TestEnvironment"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      security: []
...
  /api/v1/relay/ws:
    get:
      operationId: multiplexedRelay
      summary: Prewarm one physical socket and independently authorize subscriptions
      description: >-
        Upgrade to a WebSocket with relay protocol 1. The server sends relay_ready.
        Submit subscribe frames with a unique subscription_id, bearer token, org_id,
        silicon_ids and optional app_secret or test_key. Per-identity messages are
        wrapped as type frame with subscription_id and the ordinary v1 frame.
        Authorization failures affect only the logical subscription. The prewarm
        connection grants no access. Limit 256 subscriptions and 64 KiB client frames.
        See https://docs.hook.teamofsilicons.com/client/relay/ for heartbeats and ACKs.
      tags: [Deliveries]
      security: []
      responses:
        '101':
          description: Physical socket ready; each subscription still requires authorization
        '400':
          description: Invalid upgrade
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM opaque access token
  parameters:
...
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      schema:
        "$ref": "#/components/schemas/OrganizationId"
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 255
        pattern: "^[!-~]+$"
...
  responses:
    BadRequest:
      description: Required HTTP syntax, headers, or JSON are malformed.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
...
    Unauthorized:
      description: Credential is absent, invalid, expired, or revoked.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    Forbidden:
      description: The authenticated actor lacks authority.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
...
    NotFound:
      description: The resource does not exist or is not visible.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    Conflict:
      description: Idempotency or resource-state conflict.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
...
    UnprocessableEntity:
      description: The request violates input rules; details may explain why.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    InternalError:
      description: An unexpected internal invariant or programming failure occurred.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    ServiceUnavailable:
      description: A required dependency cannot make a safe decision.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
  schemas:
...
    OrganizationId:
      type: string
      minLength: 1
      maxLength: 100
      pattern: "^[!-~]+$"
...
    Error:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
          - code
          - message
          - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: string
    TestEnvironment:
      type: object
      required:
      - id
      - org_id
      - creator_kind
      - creator_id
      - name
      - generation
      - created_at
      - last_activity_at
      - deleted_at
      properties:
        id:
          type: string
          format: uuid
        org_id:
          "$ref": "#/components/schemas/OrganizationId"
        creator_kind:
          type: string
          enum:
          - carbon
          - silicon
        creator_id:
          type: string
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type:
          - string
          - 'null'
          maxLength: 2000
        generation:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        last_activity_at:
          type: string
          format: date-time
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
    EnvironmentWithKey:
      allOf:
      - "$ref": "#/components/schemas/TestEnvironment"
      - type: object
        required:
        - key
        - max_hooks
        properties:
          key:
            type: string
            pattern: "^[A-Za-z0-9]{32}$"
          max_hooks:
            type: integer
            const: 10
    TestIamConfiguration:
      type: object
      additionalProperties: false
      required:
      - app_id
      - app_secret
      - webhook_secret
      properties:
        app_id:
          type: string
          description: Canonical test application ID, e.g. tos>hook.
        app_secret:
          type: string
          writeOnly: true
        webhook_secret:
          type: string
          writeOnly: true
        webhook_secret_version:
          type: integer
          minimum: 1
          default: 1
    CreateEnvironment:
      type: object
      additionalProperties: false
      required:
      - name
      - iam_test_key
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type:
          - string
          - 'null'
          maxLength: 2000
        iam_test_key:
          type: string
          writeOnly: true
        iam:
          "$ref": "#/components/schemas/TestIamConfiguration"
    EnvironmentList:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/TestEnvironment"